Support retro_get_memory_{data,size}
This change adds support for get_memory methods that can be used to add native
save, rtc, and ram support to a libretro core.

For each memory type we provide a corresponding method:

- `rtc_memory(&mut self)` for `MEMORY_RTC`
- `save_memory(&mut self)` for `MEMORY_SAVE_RAM`
- `system_memory(&mut self)` for `MEMORY_SYSTEM_RAM`
- `video_memory(&mut self)` for `MEMORY_VIDEO_RAM`

A `Core` is supposed to return a mutable reference to the respective memory area
from the `*_memory` methods for each type of memory it supports. E.g. a `Core`
that supports saves should implement

```
fn save_memory( &mut self ) -> Option< &mut [u8] > {
    // return mutable reference to the save ram
}
```

The libretro documentation describes how each memory type is supposed to be
used.